+++ example/lib/main.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_defender/flutter_defender.dart';
 
 import 'src/app/example_app.dart';
 import 'src/app/session/session_controller.dart';
@@ -13,17 +12,13 @@ Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
   final SessionController sessionController = SessionController();
 
-  await FlutterDefender.instance.init(
-    otpBackgroundTimeoutSeconds: 10,
-    pinBackgroundTimeoutSeconds: 20,
-    onLogoutRequested: () {
-      sessionController.handleTimeoutLogout();
-      appNavigatorKey.currentState?.popUntil(
-        (Route<dynamic> route) => route.isFirst,
-      );
-    },
-    blockingScreenBuilder: buildExampleBlockingScreen,
-  );
+  sessionController.registerLogoutHandler(() {
+    sessionController.handleTimeoutLogout();
+    appNavigatorKey.currentState?.popUntil(
+      (Route<dynamic> route) => route.isFirst,
+    );
+  });
+  await sessionController.applyProfile(sessionController.activeProfile);
 
   runApp(
     MyApp(sessionController: sessionController, navigatorKey: appNavigatorKey),